requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
langgraph>=0.0.15
crewai>=0.28.0
autogen>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
requests>=2.31.0
pydantic>=2.4.0
fastapi>=0.104.0
uvicorn>=0.23.0
chromadb>=0.3.26
browser-use>=0.1.0
# langchain-google-genai